chore: cut out Listbox to divide responsibilities
